125_выносит анимацию из кода фрагмента в отдельный класс
