+++ app/src/main/java/com/davay/android/feature/match/presentation/MatchBottomSheetFragment.kt
@@ -1,9 +1,5 @@
 package com.davay.android.feature.match.presentation
 
-import android.animation.Animator
-import android.animation.AnimatorListenerAdapter
-import android.animation.AnimatorSet
-import android.animation.ObjectAnimator
 import android.app.Dialog
 import android.content.res.Resources
 import android.os.Bundle
@@ -12,9 +8,10 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.FrameLayout
 import androidx.lifecycle.lifecycleScope
-import com.davai.extensions.dpToPx
 import com.davay.android.databinding.FragmentMatchBottomSheetBinding
 import com.davay.android.domain.models.MovieDetails
+import com.davay.android.feature.sessionsmatched.presentation.animation.AnimationMatchDialog
+import com.davay.android.feature.sessionsmatched.presentation.animation.AnimationMatchDialogImpl
 import com.davay.android.utils.MovieDetailsHelper
 import com.davay.android.utils.MovieDetailsHelperImpl
 import com.google.android.material.bottomsheet.BottomSheetBehavior
@@ -29,6 +26,7 @@ class MatchBottomSheetFragment(private val action: (() -> Unit)? = null) :
     private val binding: FragmentMatchBottomSheetBinding
         get() = _binding!!
     private val movieDetailsHelper: MovieDetailsHelper = MovieDetailsHelperImpl()
+    private val animationMatchDialog: AnimationMatchDialog = AnimationMatchDialogImpl()
     private var movieDetails: MovieDetails? = null
     private var buttonText: String? = null
 
@@ -106,7 +104,7 @@ class MatchBottomSheetFragment(private val action: (() -> Unit)? = null) :
                 addBottomSheetCallback(object : BottomSheetBehavior.BottomSheetCallback() {
                     override fun onStateChanged(bottomSheet: View, newState: Int) {
                         if (newState == BottomSheetBehavior.STATE_EXPANDED) {
-                            animateBannerDrop()
+                            animationMatchDialog.animateBannerDrop(binding.tvBannerMatchWatch)
                         }
                     }
 
@@ -119,7 +117,7 @@ class MatchBottomSheetFragment(private val action: (() -> Unit)? = null) :
             }
 
             launchProgressButtonAnimation()
-            animateBannerDrop()
+            animationMatchDialog.animateBannerDrop(binding.tvBannerMatchWatch)
         }
     }
 
@@ -161,61 +159,15 @@ class MatchBottomSheetFragment(private val action: (() -> Unit)? = null) :
         val bottomSheet =
             dialog?.findViewById<View>(com.google.android.material.R.id.design_bottom_sheet) as? FrameLayout
                 ?: return
-
-        val width = bottomSheet.width
-        val height = bottomSheet.height
-
-        val screenWidth = Resources.getSystem().displayMetrics.widthPixels
-
-        val translationX =
-            screenWidth - width.toFloat() / 2 - bottomSheet.paddingEnd - EXTERNAL_SPACING_24_DP.dpToPx()
-        val translationY = -height.toFloat() / 2 + EXTERNAL_SPACING_24_DP.dpToPx()
-
-        val scaleX = ObjectAnimator.ofFloat(bottomSheet, View.SCALE_X, 0f)
-        val scaleY = ObjectAnimator.ofFloat(bottomSheet, View.SCALE_Y, 0f)
-        val translateX = ObjectAnimator.ofFloat(bottomSheet, View.TRANSLATION_X, translationX)
-        val translateY = ObjectAnimator.ofFloat(bottomSheet, View.TRANSLATION_Y, translationY)
-        val alpha = ObjectAnimator.ofFloat(bottomSheet, View.ALPHA, 0f)
-
-        val animatorSet = AnimatorSet().apply {
-            playTogether(scaleX, scaleY, translateX, translateY, alpha)
-            duration = MORPHING_ANIMATION_DURATION_500_MS
+        animationMatchDialog.animateDialogDismiss(
+            bottomSheet
+        ) {
+            dismiss()
+            action?.invoke()
         }
-
-        animatorSet.addListener(object : AnimatorListenerAdapter() {
-            override fun onAnimationEnd(animation: Animator) {
-                dismiss()
-                action?.invoke()
-            }
-        })
-
-        animatorSet.start()
     }
-
-    private fun animateBannerDrop() {
-        val banner = binding.tvBannerMatchWatch
-
-        val scaleX = ObjectAnimator.ofFloat(banner, View.SCALE_X, START_SCALE_X5, 1f)
-        val scaleY = ObjectAnimator.ofFloat(banner, View.SCALE_Y, START_SCALE_X5, 1f)
-        val alpha = ObjectAnimator.ofFloat(banner, View.ALPHA, 0f, 1f)
-
-        scaleX.duration = BANNER_DROP_DURATION_500_MS
-        scaleY.duration = BANNER_DROP_DURATION_500_MS
-        alpha.duration = BANNER_DROP_DURATION_500_MS
-
-        val animatorSet = AnimatorSet().apply {
-            playTogether(scaleX, scaleY, alpha)
-        }
-
-        animatorSet.start()
-    }
-
 
     companion object {
-        private const val EXTERNAL_SPACING_24_DP = 24
-        private const val MORPHING_ANIMATION_DURATION_500_MS = 500L
-        private const val BANNER_DROP_DURATION_500_MS = 500L
-        private const val START_SCALE_X5 = 5f
         private const val ARG_MOVIE_DETAILS = "movie_details"
         private const val ARG_BUTTON_TEXT = "button_text"
