+++ app/src/main/java/com/davay/android/feature/sessionsmatched/presentation/animation/AnimationMatchDialog.kt
@@ -1,0 +1,12 @@
+package com.davay.android.feature.sessionsmatched.presentation.animation
+
+import android.view.View
+
+interface AnimationMatchDialog {
+    fun animateDialogDismiss(
+        bottomSheet: View,
+        onAnimationEnd: () -> Unit
+    )
+
+    fun animateBannerDrop(banner: View)
+}
